+++ Vault-Coder/src/utopia/vault/coder/controller/writer/EnumerationWriter.scala
@@ -26,36 +26,42 @@ object EnumerationWriter
 	def apply(enum: Enum)(implicit setup: ProjectSetup, codec: Codec) =
 	{
 		// Enumeration doesn't need to be imported in its own file
-		val enumDataType = ScalaType.basic(enum.name.singular)
+		val enumDataType = ScalaType.basic(enum.name)
 		
 		File(enum.packagePath,
 			// Writes the enumeration trait first
-			TraitDeclaration(enum.name.singular,
+			TraitDeclaration(enum.name,
 				// Each value contains an id so that it can be referred from the database
-				properties = Vector(PropertyDeclaration.newAbstract("id", ScalaType.int)),
+				properties = Vector(PropertyDeclaration.newAbstract("id", ScalaType.int,
+					description = "Id used for this value in database / SQL")),
 				isSealed = true
 			),
 			// Enumeration values are nested within a companion object
-			ObjectDeclaration(enum.name.singular,
+			ObjectDeclaration(enum.name,
 				// Contains the .values -property
 				properties = Vector(
 					ImmutableValue("values", enum.values.map { value =>
-						Reference(s"${enum.packagePath}.${enum.name}", value.singular) }.toSet,
-						explicitOutputType = Some(ScalaType.vector(enumDataType)))(
-						s"Vector(${enum.values.map { _.singular }.mkString(", ")})")
+						Reference(s"${enum.packagePath}.${enum.name}", value) }.toSet,
+						explicitOutputType = Some(ScalaType.vector(enumDataType)),
+						description = "All available values of this enumeration")(
+						s"Vector(${enum.values.mkString(", ")})")
 				),
 				// Contains an id to enum value -function (one with Try, another with Option)
 				methods = Set(
-					MethodDeclaration("findForId")(Parameter("id", ScalaType.int))("values.find { _.id == id }"),
+					MethodDeclaration("findForId",
+						returnDescription = s"${enum.name} matching that id. None if the id didn't match any ${enum.name}")(
+						Parameter("id", ScalaType.int, description = s"Id representing a ${enum.name}"))(
+						"values.find { _.id == id }"),
 					MethodDeclaration("forId",
-						codeReferences = Set(Reference.collectionExtensions, Reference.noSuchElementException))(
-						Parameter("id", ScalaType.int))(
+						codeReferences = Set(Reference.collectionExtensions, Reference.noSuchElementException),
+						returnDescription = s"${enum.name} matching that id. Failure if no suitable value was found.")(
+						Parameter("id", ScalaType.int, description = s"Id matching a ${enum.name}"))(
 						s"findForId(id).toTry { new NoSuchElementException(${
 							s"No value of ${enum.name} matches id '${"s${id}"}'".quoted}) }")
 				),
 				// Contains an object for each value
 				nested = enum.values.zipWithIndex.map { case (valueName, index) =>
-					ObjectDeclaration(valueName.singular, Vector(Extension(enumDataType)),
+					ObjectDeclaration(valueName, Vector(Extension(enumDataType)),
 						// The objects don't contain other properties except for 'id'
 						properties = Vector(ImmutableValue("id", isOverridden = true)(s"${index + 1}")),
 						isCaseObject = true